ATTENTION: BUG IN SCHEDULER
Please note there is a bug currently in the scheduler. Unless you briefly run a firing you wish to schedule, BEFORE scheduling it, the previous firing curve last run is scheduled. So ... start and stop the curve you want to schedule, then refresh the browser and schedule it in. You can verify the correct firing curve will run by refreshing the screen after scheduling the firing. THis will be corrected in the coming days/week.